+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Description: Flexiv Rizon4 (MJCF)
 - Description: Flexiv Rizon4 (Xacro)
 - Description: Robotiq 2F-85 v4 (URDF) (thanks to @nickswalker)
+- Description: SO-ARM101 Parallel Gripper (URDF)
 - Description: Franka descriptions added XACRO_ARGS_NO_HAND property (thanks to @nickswalker)
 - Loaders: Allow overriding `XACRO_ARGS` via `load_robot_description` (thanks to @nickswalker)